增加节点子辈映射表生成
