+++ renderer/lib/game/game/save/rollback.ts
@@ -1,4 +1,5 @@
 import {LogicAction} from "@lib/game/game/logicAction";
+import {Game} from "@lib/game/game/game";
 
 export enum NodeType {
     TreeNode = "TreeNode",
@@ -40,7 +41,34 @@ export type ContentNodeData = {
 }
 
 export class ContentNode<T = any> extends Node<T> {
-    child: RenderableNode | null;
+    static forEachParent(node: RenderableNode, callback: (node: RenderableNode) => void) {
+        const seen: Set<RenderableNode> = new Set();
+        let current = node;
+        while (current) {
+            if (seen.has(current)) {
+                break;
+            }
+            seen.add(current);
+            callback(current);
+            current = current.parent;
+        }
+    }
+
+    static forEachChild(node: RenderableNode, callback: (node: RenderableNode) => void) {
+        const seen: Set<RenderableNode> = new Set();
+        let current = node;
+        while (current) {
+            if (seen.has(current)) {
+                break;
+            }
+            seen.add(current);
+            callback(current);
+            current = current.child;
+        }
+    }
+
+    child?: RenderableNode | null;
+    initChild?: RenderableNode | null;
     parent: RenderableNode | null;
     action: LogicAction.Actions;
 
@@ -50,7 +78,7 @@ export class ContentNode<T = any> extends Node<T> {
         parent?: RenderableNode | null,
         callee?: LogicAction.Actions
     ) {
-        super(id, NodeType.ContentNode);
+        super(Game.getIdManager().prefix("node", id, "-"), NodeType.ContentNode);
         this.child = child || null;
         this.parent = parent || null;
         this.action = callee
@@ -76,9 +104,19 @@ export class ContentNode<T = any> extends Node<T> {
     }
 
     /**
-     * For chaining
+     * To track the changes of the child
+     * should only be called when constructing the tree
      */
-    addChild(child: RenderableNode) {
+    setInitChild(child: RenderableNode) {
+        this.initChild = child;
+        return this.setChild(child);
+    }
+
+    /**
+     * Public method for setting the content of the node
+     * should only be called when changing the state in-game
+     */
+    public addChild(child: RenderableNode) {
         this.setChild(child);
         return this;
     }
@@ -146,17 +184,21 @@ export class RootNode extends ContentNode {
 
     forEach(callback: (node: RenderableNode) => void) {
         const queue = [this.child];
+        const seen: Set<RenderableNode> = new Set();
         while (queue.length > 0) {
             const node = queue.shift();
             if (!node) {
                 continue;
             }
+            if (seen.has(node)) {
+                continue;
+            }
+            seen.add(node);
             callback(node);
             if (node instanceof ContentNode) {
                 queue.push(node.child);
             }
         }
-        return void 0;
     }
 }
